Add "dev_exe" variant to build.py
This mode will be for a standalone executable (without overlays) with
some overlay actors embedded in and it will boot into some stage 
directly (not main menu). In the future it will include some new 
"homebrew" debug actors (for example debug camera). This is not
a matching executable and it's meant to be freely modified, so checks 
are disabled in this variant.