Embed an overlay, force load specific stage in dev variant
Make minimal changes to the code (only in dev variant) to make it
possible to use the built executable as a standalone executable
(no overlays) and force load a specific stage (not main menu).

Create a new "contrib/" directory which will be a (separate!) home for
new (debug) code we wish to add - not present in the original game.